:tractor: :sparkles: Split UI out, add cobra command
