+++ cmd/chat/main.go
@@ -1,241 +1,15 @@
 package main
 
 import (
-	"fmt"
-	"github.com/charmbracelet/bubbles/key"
-	"github.com/charmbracelet/bubbles/textarea"
+	context2 "context"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 	"github.com/go-go-golems/geppetto/pkg/context"
-	"github.com/muesli/reflow/wordwrap"
-	"github.com/rs/zerolog/log"
+	"github.com/go-go-golems/geppetto/pkg/steps/openai"
+	"github.com/go-go-golems/glazed/pkg/cli"
+	"github.com/go-go-golems/glazed/pkg/cmds"
+	"github.com/go-go-golems/glazed/pkg/cmds/layers"
+	"github.com/spf13/cobra"
 )
-
-type KeyMap struct {
-	SelectPrevMessage key.Binding
-	SelectNextMessage key.Binding
-	UnfocusMessage    key.Binding
-	FocusMessage      key.Binding
-	SubmitMessage     key.Binding
-	ScrollUp          key.Binding
-	ScrollDown        key.Binding
-	Quit              key.Binding
-}
-
-var DefaultKeyMap = KeyMap{
-	SelectPrevMessage: key.NewBinding(key.WithKeys("up")),
-	SelectNextMessage: key.NewBinding(key.WithKeys("down")),
-	UnfocusMessage:    key.NewBinding(key.WithKeys("esc", "ctrl+g")),
-	FocusMessage:      key.NewBinding(key.WithKeys("enter")),
-	SubmitMessage:     key.NewBinding(key.WithKeys("shift+enter")),
-	ScrollUp:          key.NewBinding(key.WithKeys("shift+pgup")),
-	ScrollDown:        key.NewBinding(key.WithKeys("shift+pgdown")),
-	Quit:              key.NewBinding(key.WithKeys("ctrl+c")),
-}
-
-type Style struct {
-	UnselectedMessage lipgloss.Style
-	SelectedMessage   lipgloss.Style
-	FocusedMessage    lipgloss.Style
-}
-
-type BorderColors struct {
-	Unselected string
-	Selected   string
-	Focused    string
-}
-
-func DefaultStyles() *Style {
-	lightModeColors := BorderColors{
-		Unselected: "#CCCCCC",
-		Selected:   "#FFB6C1", // Light pink
-		Focused:    "#FFFF99", // Light yellow
-	}
-
-	darkModeColors := BorderColors{
-		Unselected: "#444444",
-		Selected:   "#DD7090", // Desaturated pink for dark mode
-		Focused:    "#DDDD77", // Desaturated yellow for dark mode
-	}
-
-	return &Style{
-		UnselectedMessage: lipgloss.NewStyle().Border(lipgloss.NormalBorder()).
-			Padding(1, 1).
-			BorderForeground(lipgloss.AdaptiveColor{
-				Light: lightModeColors.Unselected,
-				Dark:  darkModeColors.Unselected,
-			}),
-		SelectedMessage: lipgloss.NewStyle().Border(lipgloss.ThickBorder()).
-			Padding(1, 1).
-			BorderForeground(lipgloss.AdaptiveColor{
-				Light: lightModeColors.Selected,
-				Dark:  darkModeColors.Selected,
-			}),
-		FocusedMessage: lipgloss.NewStyle().Border(lipgloss.NormalBorder()).
-			Padding(1, 1).
-			BorderForeground(lipgloss.AdaptiveColor{
-				Light: lightModeColors.Focused,
-				Dark:  darkModeColors.Focused,
-			}),
-	}
-}
-
-type errMsg error
-
-type model struct {
-	contextManager *context.Manager
-	// not really what we want, but use this for now, we'll have to either find a normal text box,
-	// or implement wrapping ourselves.
-	textArea textarea.Model
-	// is the textarea currently focused
-	focused bool
-	// currently selected message, always valid
-	selectedIdx int
-	err         error
-	keyMap      KeyMap
-
-	style  *Style
-	width  int
-	height int
-}
-
-func (m *model) updateKeyBindings() {
-	if m.focused {
-		m.keyMap.SelectNextMessage.SetEnabled(false)
-		m.keyMap.SelectPrevMessage.SetEnabled(false)
-		m.keyMap.FocusMessage.SetEnabled(false)
-		m.keyMap.UnfocusMessage.SetEnabled(true)
-	} else {
-		m.keyMap.SelectNextMessage.SetEnabled(true)
-		m.keyMap.SelectPrevMessage.SetEnabled(true)
-		m.keyMap.FocusMessage.SetEnabled(true)
-		m.keyMap.UnfocusMessage.SetEnabled(false)
-	}
-}
-
-func initialModel(manager *context.Manager) model {
-	ret := model{
-		contextManager: manager,
-		style:          DefaultStyles(),
-		keyMap:         DefaultKeyMap,
-	}
-
-	ret.textArea = textarea.New()
-	ret.textArea.Placeholder = "Once upon a time..."
-	ret.textArea.Focus()
-	ret.focused = true
-
-	ret.selectedIdx = len(ret.contextManager.GetMessages()) - 1
-
-	ret.updateKeyBindings()
-
-	return ret
-}
-
-func (m model) Init() tea.Cmd {
-	return textarea.Blink
-}
-
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var cmds []tea.Cmd
-	var cmd tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, m.keyMap.UnfocusMessage):
-			// we need to access with the pointer
-			if m.focused {
-				m.textArea.Blur()
-				m.focused = false
-				m.updateKeyBindings()
-			}
-		case key.Matches(msg, m.keyMap.Quit):
-			return m, tea.Quit
-
-		case key.Matches(msg, m.keyMap.FocusMessage):
-			if !m.focused {
-				cmd = m.textArea.Focus()
-
-				m.focused = true
-				m.updateKeyBindings()
-			}
-
-		case key.Matches(msg, m.keyMap.SelectNextMessage):
-			if m.selectedIdx < len(m.contextManager.GetMessages())-1 {
-				m.selectedIdx++
-			}
-
-		case key.Matches(msg, m.keyMap.SelectPrevMessage):
-			if m.selectedIdx > 0 {
-				m.selectedIdx--
-			}
-
-		case key.Matches(msg, m.keyMap.SubmitMessage):
-			if m.focused {
-				// XXX actually send the whole context to the LLM
-			}
-
-		default:
-			if m.focused {
-				m.textArea, cmd = m.textArea.Update(msg)
-				cmds = append(cmds, cmd)
-			}
-		}
-
-	case tea.WindowSizeMsg:
-		h, _ := m.style.SelectedMessage.GetFrameSize()
-		newWidth := msg.Width - h
-		m.textArea.SetWidth(newWidth)
-		m.width = msg.Width
-		m.height = msg.Height
-
-	// We handle errors just like any other message
-	case errMsg:
-		m.err = msg
-		return m, nil
-
-	default:
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
-func (m model) View() string {
-	ret := ""
-
-	for idx := range m.contextManager.GetMessages() {
-		v := m.contextManager.GetMessages()[idx].Text
-
-		w, _ := m.style.SelectedMessage.GetFrameSize()
-
-		w_ := wordwrap.NewWriter(m.width - w)
-		_, err := fmt.Fprintf(w_, v)
-		if err != nil {
-			panic(err)
-		}
-		v = w_.String()
-		if idx == m.selectedIdx && !m.focused {
-			v = m.style.SelectedMessage.Render(v)
-		} else {
-			v = m.style.UnselectedMessage.Render(v)
-		}
-		ret += v
-		ret += "\n"
-	}
-
-	v := m.textArea.View()
-	if m.focused {
-		v = m.style.FocusedMessage.Render(v)
-	} else {
-		v = m.style.UnselectedMessage.Render(v)
-	}
-
-	ret += v
-	ret += "\n"
-
-	return ret
-}
 
 const veryLongLoremIpsum = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec a diam lectus. " +
 	"Sed sit amet ipsum mauris. Maecenas congue ligula ac quam viverra nec consectetur ante hendrerit. " +
@@ -247,7 +21,32 @@ const veryLongLoremIpsum = "Lorem ipsum dolor sit amet, consectetur adipiscing e
 	"Curabitur dapibus enim sit amet elit pharetra tincidunt feugiat nisl imperdiet. " +
 	"Ut convallis libero in urna ultrices accumsan. Donec sed odio eros."
 
-func main() {
+type ChatCommand struct {
+	*cmds.CommandDescription
+}
+
+func NewChatCommand() (*ChatCommand, error) {
+	openaiParameterLayer, err := openai.NewClientParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+
+	return &ChatCommand{
+		CommandDescription: cmds.NewCommandDescription(
+			"chat",
+			cmds.WithShort("chat with the mechanical god in the clouds"),
+			cmds.WithLayers(
+				openaiParameterLayer,
+			),
+		),
+	}, nil
+}
+
+func (c *ChatCommand) Run(
+	ctx context2.Context,
+	parsedLayers map[string]*layers.ParsedParameterLayer,
+	ps map[string]interface{},
+) error {
 	messages := []*context.Message{
 		// different substrings of veryLongLoremIpsum
 		{
@@ -268,6 +67,24 @@ func main() {
 	p := tea.NewProgram(initialModel(ctxtManager))
 
 	if _, err := p.Run(); err != nil {
-		log.Fatal().Err(err).Msg("Error while handling bubbletea")
+		return err
 	}
+
+	return nil
+}
+
+var RootCmd = &cobra.Command{
+	Use:   "chat",
+	Short: "Chat with the mechanical god in the clouds",
+}
+
+func main() {
+	chatCmd, err := NewChatCommand()
+	cobra.CheckErr(err)
+
+	chatCobraCommand, err := cli.BuildCobraCommandFromBareCommand(chatCmd)
+	cobra.CheckErr(err)
+
+	err = chatCobraCommand.Execute()
+	cobra.CheckErr(err)
 }
